Fix issue #357. Interrupt is triggered when Transmit is completed, calling the receiver function
When transmitting a packet, the GDO0 pin is set to trigger when the transmission is ongoing.
Unfortunately this will call the user interrupt even if the user intended to use only on receiving packets.

The issue occurs because GDO0 is monitored for transmit completion in the transmit() function, but a
change will trigger the user interrupt as well. The solution is to redirect the interrupt to a separate
interrupt handler until the transmit function is completed, then restoring the original handler before
returning.