[CC1101] Data shaping and encoding params in right order
For some reason, the `RADIOLIB_SHAPING_NONE` was being passed to `setEncoding` and `RADIOLIB_ENCODING_NRZ` was being passed to `setDataShaping`. Flipped these to be passed to the correct functions.

Note: These are both defined as `0x00` so this should have no impact on functionality, just makes initialization less confusing.