Module: Mark some functions as virtual
In an effort to support RadioLib on Tock (github.com/tock/tock) we want
to replace some of the Module tranfer functions. This is because we want
to replace the entire read/write command with an operating system call,
instead of performing individual options. For example we don't want to
manually control the CS line, but instead want to just specify a read
buffer, write buffer and length and have Tock perform the operation for
us.

The current Hal abstraction isn't powerful enough for this, so we need
to override some Module functions instead.

Signed-off-by: Alistair Francis <alistair@alistair23.me>