NonArduino: Tock: Improve the reliability of delays
Using the Tock libtocksync_alarm_delay_ms() syscall has too much
overhead, both in terms of extra time but also jitter. This means it's
not a reliable source for short but accurate delays, such as the kind
used in LoRaWAN.

This patch instead uses a busy loop for short (less then 5 second)
delays. This might have some impact on performance and power as we are
busy running in a loop, but overall we end up with a much more accurate
time and working LoRaWAN.

Signed-off-by: Alistair Francis <alistair@alistair23.me>